# packages/core-schemas/openapi.yaml
openapi: 3.1.0
info:
  title: DreamRight API
  description: |
    AI-powered webtoon and short-form drama production API.

    ## Authentication
    All endpoints require Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-api-token>
    ```

    ## Workflow
    The typical workflow for creating a webtoon project is:
    1. Create a project (`POST /projects`)
    2. Expand a story prompt (`POST /projects/{id}/story`)
    3. Generate character assets (`POST /projects/{id}/characters/{id}/assets`)
    4. Generate location assets (`POST /projects/{id}/locations/{id}/assets`)
    5. Generate chapters sequentially (`POST /projects/{id}/chapters`)
    6. Generate panels for each chapter (`POST /projects/{id}/chapters/{id}/panels`)

    ## Asset Dependencies
    - Panel generation requires character portraits and location references to exist
    - Chapter N generation requires Chapter N-1 to exist (for story continuity)
    - Panel generation for Chapter N requires Chapter N-1 panels to exist (for visual continuity)

    ## Async Operations
    Long-running generation operations return `202 Accepted` with a job resource.
    Poll the job status endpoint or use webhooks for completion notifications.
  version: 1.0.0
  contact:
    name: DreamRight Support
  license:
    name: MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
security:
  - bearerAuth: []
...
paths:
  /projects:
    get:
      tags: [Projects]
      summary: List all projects
      operationId: listProjects
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ProjectStatus'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProjectList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        default:
          $ref: '#/components/responses/InternalError'

    post:
      tags: [Projects]
      summary: Create a new project
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
            example:
              name: "My Webtoon"
              format: "webtoon"
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        default:
          $ref: '#/components/responses/InternalError'
  /projects/{projectId}:
    parameters:
      - $ref: '#/components/parameters/projectId'

    get:
      tags: [Projects]
      summary: Get project details
      operationId: getProject
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalError'

    patch:
      tags: [Projects]
      summary: Update project
      operationId: updateProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: Project updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [Projects]
      summary: Delete project
      operationId: deleteProject
      responses:
        '204':
          description: Project deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalError'
  /projects/{projectId}/status:
    parameters:
      - $ref: '#/components/parameters/projectId'

    get:
      tags: [Projects]
      summary: Get project generation status
      description: Returns detailed status of all assets and generation progress
      operationId: getProjectStatus
      responses:
        '200':
          description: Project status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectStatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalError'

  # Story endpoints - Resource-oriented: POST creates/expands the story
...
  /projects/{projectId}/chapters/{chapterId}:
    parameters:
      - $ref: '#/components/parameters/projectId'
      - $ref: '#/components/parameters/chapterId'

    get:
      tags: [Chapters]
      summary: Get chapter details
      operationId: getChapter
      responses:
        '200':
          description: Chapter details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chapter'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [Chapters]
      summary: Delete chapter
      operationId: deleteChapter
      responses:
        '204':
          description: Chapter deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalError'

  # Panel endpoints - Resource-oriented
  /projects/{projectId}/chapters/{chapterId}/panels:
    parameters:
      - $ref: '#/components/parameters/projectId'
      - $ref: '#/components/parameters/chapterId'

    get:
      tags: [Panels]
      summary: List panels for a chapter
      operationId: listPanels
      parameters:
        - name: scene_number
          in: query
          description: Filter by scene number
          schema:
            type: integer
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        '200':
          description: List of panels
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPanelList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalError'

    post:
      tags: [Panels]
      summary: Generate panels for chapter
      description: |
        Generates panel images for all scenes in a chapter using AI.

        **Dependencies**:
        - All characters referenced in panels must have portrait assets
        - All locations referenced in scenes must have reference assets
        - Chapter N-1 must exist (for story context)
        - Chapter N-1 panels should exist (for visual continuity across chapters)

        Panels are generated sequentially within each scene to maintain visual continuity.

        This is an async operation.
      operationId: createPanels
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePanelsRequest'
      responses:
        '202':
          description: Panel generation started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchJobResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Dependencies not met
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DependencyError'
        '429':
          $ref: '#/components/responses/RateLimited'
        default:
          $ref: '#/components/responses/InternalError'
  /projects/{projectId}/chapters/{chapterId}/scenes/{sceneNumber}/panels:
    parameters:
      - $ref: '#/components/parameters/projectId'
      - $ref: '#/components/parameters/chapterId'
      - $ref: '#/components/parameters/sceneNumber'

    post:
      tags: [Panels]
      summary: Generate panels for a specific scene
      description: Generate panel images for a single scene. Async operation.
      operationId: createScenePanels
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScenePanelsRequest'
      responses:
        '202':
          description: Panel generation started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchJobResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Dependencies not met
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DependencyError'
        '429':
          $ref: '#/components/responses/RateLimited'
        default:
          $ref: '#/components/responses/InternalError'

  # Asset file endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  parameters:
    projectId:
      name: projectId
      in: path
      required: true
      description: Project ID
      schema:
        type: string
...
    chapterId:
      name: chapterId
      in: path
      required: true
      description: Chapter ID (e.g., ch1, ch2)
      schema:
        type: string
    sceneNumber:
      name: sceneNumber
      in: path
      required: true
      description: Scene number (1-indexed)
      schema:
        type: integer
        minimum: 1
...
    limit:
      name: limit
      in: query
      description: Maximum number of items to return
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    offset:
      name: offset
      in: query
      description: Number of items to skip
      schema:
        type: integer
        default: 0
        minimum: 0
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "unauthorized"
            message: "Authentication required"
    Forbidden:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "forbidden"
            message: "You do not have access to this resource"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "rate_limited"
            message: "Rate limit exceeded. Try again in 60 seconds."
      headers:
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    ProjectFormat:
      type: string
      enum: [webtoon, short_drama]
    ProjectStatus:
      type: string
      enum: [draft, in_progress, completed]
    Genre:
      type: string
      enum: [romance, action, fantasy, thriller, slice_of_life, horror, comedy, drama, mystery, scifi]
...
    ChapterStatus:
      type: string
      enum: [outlined, generating, completed]
    JobStatus:
      type: string
      enum: [pending, running, completed, failed, cancelled]
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    DependencyError:
      type: object
      required: [code, message, missing_dependencies]
      properties:
        code:
          type: string
          enum: [dependency_not_met]
        message:
          type: string
        missing_dependencies:
          type: array
          items:
            $ref: '#/components/schemas/MissingDependency'
    MissingDependency:
      type: object
      required: [type, id, name, resolution]
      properties:
        type:
          $ref: '#/components/schemas/DependencyType'
        id:
          type: string
          description: Resource ID
        name:
          type: string
          description: Human-readable name
        resolution:
          type: string
          description: How to resolve this dependency
          example: "Generate character portrait: POST /projects/{id}/characters/{id}/assets"
...
    PaginationMeta:
      type: object
      required: [total, limit, offset]
      properties:
        total:
          type: integer
          description: Total number of items
        limit:
          type: integer
        offset:
          type: integer
        has_more:
          type: boolean

    # Paginated lists
    PaginatedProjectList:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProjectSummary'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginatedPanelList:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Panel'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    BatchJobResponse:
      type: object
      required: [job_id, status, poll_url, total_items]
      properties:
        job_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/JobStatus'
        poll_url:
          type: string
          format: uri
        total_items:
          type: integer
          description: Number of items to process

    # Project schemas
    ProjectSummary:
      type: object
      required: [id, name, format, status, created_at]
      properties:
        id:
          type: string
        name:
          type: string
        format:
          $ref: '#/components/schemas/ProjectFormat'
        status:
          $ref: '#/components/schemas/ProjectStatus'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        story_title:
          type: string
          nullable: true
        character_count:
          type: integer
        location_count:
          type: integer
        chapter_count:
          type: integer
    Project:
      type: object
      required: [id, name, format, status, created_at, characters, locations, chapters]
      properties:
        id:
          type: string
        name:
          type: string
        format:
          $ref: '#/components/schemas/ProjectFormat'
        status:
          $ref: '#/components/schemas/ProjectStatus'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        story:
          $ref: '#/components/schemas/Story'
          nullable: true
        characters:
          type: array
          items:
            $ref: '#/components/schemas/Character'
        locations:
          type: array
          items:
            $ref: '#/components/schemas/Location'
        chapters:
          type: array
          items:
            $ref: '#/components/schemas/ChapterSummary'
        original_prompt:
          type: string
    CreateProjectRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        format:
          $ref: '#/components/schemas/ProjectFormat'
          default: webtoon
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        status:
          $ref: '#/components/schemas/ProjectStatus'
    ProjectStatusResponse:
      type: object
      required: [project_id, status, story_expanded, characters, locations, chapters]
      properties:
        project_id:
          type: string
        status:
          $ref: '#/components/schemas/ProjectStatus'
        story_expanded:
          type: boolean
        characters:
          type: object
          required: [total, with_assets, missing_assets]
          properties:
            total:
              type: integer
            with_assets:
              type: integer
            missing_assets:
              type: array
              items:
                $ref: '#/components/schemas/AssetStatusItem'
        locations:
          type: object
          required: [total, with_assets, missing_assets]
          properties:
            total:
              type: integer
            with_assets:
              type: integer
            missing_assets:
              type: array
              items:
                $ref: '#/components/schemas/AssetStatusItem'
        chapters:
          type: object
          required: [total_beats, generated, with_panels]
          properties:
            total_beats:
              type: integer
            generated:
              type: integer
            with_panels:
              type: integer
            missing_chapters:
              type: array
              items:
                type: object
                required: [beat_number, beat_name]
                properties:
                  beat_number:
                    type: integer
                  beat_name:
                    type: string
        panels:
          type: object
          required: [total, generated]
          properties:
            total:
              type: integer
            generated:
              type: integer
    AssetStatusItem:
      type: object
      required: [id, name, asset_type]
      properties:
        id:
          type: string
        name:
          type: string
        asset_type:
          type: string
          enum: [portrait, reference, panel]

    # Story schemas
...
    Story:
      type: object
      required: [id, title, genre, tone, episode_count]
      properties:
        id:
          type: string
        title:
          type: string
        logline:
          type: string
        genre:
          $ref: '#/components/schemas/Genre'
        tone:
          $ref: '#/components/schemas/Tone'
        themes:
          type: array
          items:
            type: string
        target_audience:
          type: string
        episode_count:
          type: integer
        synopsis:
          type: string
        story_beats:
          type: array
          items:
            $ref: '#/components/schemas/StoryBeat'
...
    Character:
      type: object
      required: [id, name, role]
      properties:
        id:
          type: string
        name:
          type: string
        role:
          $ref: '#/components/schemas/CharacterRole'
        age:
          type: string
        description:
          $ref: '#/components/schemas/CharacterDescription'
        visual_tags:
          type: array
          items:
            type: string
        assets:
          $ref: '#/components/schemas/CharacterAssets'
...
    Location:
      type: object
      required: [id, name, type]
      properties:
        id:
          type: string
        name:
          type: string
        type:
          $ref: '#/components/schemas/LocationType'
        description:
          type: string
        visual_tags:
          type: array
          items:
            type: string
        assets:
          $ref: '#/components/schemas/LocationAssets'
...
    ChapterSummary:
      type: object
      required: [id, number, title, status]
      properties:
        id:
          type: string
        number:
          type: integer
        title:
          type: string
        status:
          $ref: '#/components/schemas/ChapterStatus'
        scene_count:
          type: integer
        panel_count:
          type: integer
        generated_panel_count:
          type: integer
    Chapter:
      type: object
      required: [id, number, title, status, scenes]
      properties:
        id:
          type: string
        number:
          type: integer
        title:
          type: string
        summary:
          type: string
        status:
          $ref: '#/components/schemas/ChapterStatus'
        scenes:
          type: array
          items:
            $ref: '#/components/schemas/Scene'
...
    Scene:
      type: object
      required: [id, number, panels]
      properties:
        id:
          type: string
        number:
          type: integer
        location_id:
          type: string
          nullable: true
        time_of_day:
          $ref: '#/components/schemas/TimeOfDay'
        weather:
          type: string
        character_ids:
          type: array
          items:
            type: string
        description:
          type: string
        mood:
          type: string
        panels:
          type: array
          items:
            $ref: '#/components/schemas/Panel'
        continues_from_previous_chapter:
          type: boolean

    # Panel schemas
...
    CreatePanelsRequest:
      type: object
      properties:
        style:
          type: string
          default: webtoon
          description: Art style for generation
    CreateScenePanelsRequest:
      type: object
      properties:
        style:
          type: string
          default: webtoon
          description: Art style for generation

    # Asset schemas
...